wrap CommentsContainer in new App component #9175
